Don't bail on network errors
If CircleCI has a timeout or you disconnect from the network, just sleep and
retry again in a little bit.

Adds tests that the HTTP logic works as expected. I'd rather not go over the
network to test this, but actually reconstructing the Go HTTP logic seems like
overkill, and I doubt the test suite for this will ever be particularly large.